% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare-multi.R
\name{compare_multi_risk}
\alias{compare_multi_risk}
\title{Compare probability of being under risk constraints}
\usage{
compare_multi_risk(object_list, object_names,
  figure_dir = "compare_figure/")
}
\arguments{
\item{object_list}{list of mcmc results}

\item{object_names}{list of model names}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Compare probability of being under risk constraints
}
