% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare-multi.R
\name{compare_multi_risk_wBias}
\alias{compare_multi_risk_wBias}
\title{Add bias to recalculate risk probabilities}
\usage{
compare_multi_risk_wBias(object_list, object_names,
  figure_dir = "compare_figure/")
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Add bias to recalculate risk probabilities
}
