% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-extract.R
\name{do_extract}
\alias{do_extract}
\title{Extract model fits and load into lsdOutput object class for plotting}
\usage{
do_extract(
  dir = ".",
  data = TRUE,
  map = FALSE,
  mcmc = FALSE,
  variational = FALSE
)
}
\arguments{
\item{dir}{the directory to extract from}

\item{data}{extract a dat file}

\item{map}{extract an MAP file}

\item{mcmc}{extract one or more MCMC files}

\item{variational}{extract a var file}
}
\description{
Extract model fits and load into lsdOutput object class for plotting
}
