% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-extract_old.R
\name{do_extract_old}
\alias{do_extract_old}
\title{Extract model fits and load into lsdOutput object class for plotting}
\usage{
do_extract_old(dir = ".", data = TRUE, map = FALSE, mcmc = FALSE,
  variational = FALSE, model = character())
}
\arguments{
\item{dir}{the directory to extract from}

\item{data}{extract a dat file}

\item{map}{extract an MAP file}

\item{mcmc}{extract one or more MCMC files}

\item{variational}{extract a var file}

\item{model}{a name or label for the model being extracted}
}
\description{
Extract model fits and load into lsdOutput object class for plotting
}
