% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-plot.R
\name{do_plot}
\alias{do_plot}
\title{Plot model fits}
\usage{
do_plot(object, map = FALSE, mcmc = FALSE, figure_dir = "figure/")
}
\arguments{
\item{object}{the lsd object}

\item{map}{plot MAP if a .map output is available}

\item{mcmc}{plot MCMC if a .mcmc output is available}

\item{figure_dir}{the directory to save figures to}
}
\description{
This is a wrapper that generates most of the LSD model plots.
}
