% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-plot-compare.R
\name{do_plot_compare}
\alias{do_plot_compare}
\title{Plot model fits}
\usage{
do_plot_compare(
  object_list,
  object_names,
  single_stock = TRUE,
  figure_dir = "compare_figure/"
)
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{single_stock}{TRUE for comparing models within a single stock assessment, FALSE for comparing multiple stocks and models (e.g. reference points study)}

\item{figure_dir}{the directory to save figures to}
}
\description{
Plot model fits
}
