% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-list.R
\docType{methods}
\name{extract,list-method}
\alias{extract,list-method}
\title{Extract thingy}
\usage{
\S4method{extract}{list}(object, pars, permuted = TRUE, include = TRUE)
}
\arguments{
\item{object}{list object produced by read_stan_mcmc}

\item{pars}{character vector of parameters to be extracted (or not)}

\item{permuted}{should the MCMC sample chains be permuted and combined?}

\item{include}{should pars be extracted (TRUE) or everything except the pars (FALSE)}
}
\description{
Extract thingy
}
