% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-list.R
\name{extract}
\alias{extract}
\title{Extract MCMC samples}
\usage{
extract(object, ...)
}
\arguments{
\item{object}{list object passed to rstan extract}

\item{...}{other things passed to rstan extract}
}
\description{
Extract MCMC samples from a list produced by \code{read_stan_mcmc}
}
