% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsdOutput-class.R
\docType{methods}
\name{initialize,lsdOutput-method}
\alias{initialize,lsdOutput-method}
\title{Set up the object class}
\usage{
\S4method{initialize}{lsdOutput}(.Object, model.name)
}
\arguments{
\item{.Object}{no idea}

\item{model.name}{a model name}
}
\description{
Set up the object class
}
