% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{looic}
\alias{looic}
\title{Table computing leave-one-out information criterion for various datasets}
\usage{
looic(object_list, object_names, figure_dir = "compare_figure/")
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{figure_dir}{the directory to save to}
}
\description{
Table computing leave-one-out information criterion for various datasets
}
