% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsdSave.R
\name{lsdSave}
\alias{lsdSave}
\title{Save lsdOutput object}
\usage{
lsdSave(object, save.dir = ".")
}
\arguments{
\item{object}{an LSD object}

\item{save.dir}{the directory to save to}
}
\description{
Save lsdOutput object
}
