% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass.R
\name{plot_biomass}
\alias{plot_biomass}
\title{Plot biomass measures}
\usage{
plot_biomass(
  object,
  scales = "free_x",
  show_map = TRUE,
  show_mcmc = TRUE,
  xlab = "Fishing year",
  figure_dir = "figure/",
  ref = c("Bmsy", "Bref")
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_map}{show MAP or not}

\item{show_mcmc}{show MCMC or not}

\item{xlab}{the x axis label}

\item{figure_dir}{the directory to save to}

\item{ref}{specify Bmsy, Bref, or both}
}
\description{
Plots three types of biomass.
}
