% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catch.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot catch}
\usage{
plot_catch(object, show_proj = FALSE, scales = "free",
  xlab = "Fishing year", ylab = "Catch (tonnes)",
  figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{show_proj}{show projection or not}

\item{scales}{free or fixed}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot the catch data and fit to the data.
}
