% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catch.R
\name{plot_catch_rule}
\alias{plot_catch_rule}
\title{Plot catch}
\usage{
plot_catch_rule(
  object,
  xlab = "Fishing year",
  ylab = "Catch (tonnes)",
  figure_dir = "figure/",
  save_plot = TRUE
)
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}

\item{save_plot}{save the plot to file}
}
\description{
Plot the catch data and fit to the data.
}
