% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-catch.R
\name{plot_catch_save}
\alias{plot_catch_save}
\title{Save catch plots}
\usage{
plot_catch_save(object, figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{figure_dir}{the directory to save to}
}
\description{
Save catch plots
}
