% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare-multi.R
\name{plot_compare_catch_bio}
\alias{plot_compare_catch_bio}
\title{Compare catch vs. biomass}
\usage{
plot_compare_catch_bio(object_list, object_names,
  figure_dir = "compare_figure/")
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Compare catch vs. biomass
}
