% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot_compare_cpue}
\alias{plot_compare_cpue}
\title{Compare CPUE}
\usage{
plot_compare_cpue(
  object_list,
  object_names,
  scales = "fixed",
  xlab = "Fishing year",
  ylab = "CPUE",
  figure_dir = "compare_figure/",
  save_plot = TRUE
)
}
\arguments{
\item{scales}{free or fixed}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}

\item{save_plot}{save the plot or return it}
}
\description{
Plot the CPUE data and fit to the data.
}
