% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot_compare_lfs}
\alias{plot_compare_lfs}
\title{Compare LFs from multiple models}
\usage{
plot_compare_lfs(
  object_list,
  object_names,
  yrs = 2015:2016,
  xlab = "Midpoint of size-class (mm)",
  ylab = "Proportion at size (mm)"
)
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}
}
\description{
Compare LFs from multiple models
}
