% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot_compare_numbers}
\alias{plot_compare_numbers}
\title{Compare Numbers at length from multiple models}
\usage{
plot_compare_numbers(
  object_list,
  object_names,
  xlab = "Size (mm)",
  ylab = "Number of individuals (thousands)",
  figure_dir
)
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{figure directory}
}
\description{
Compare Numbers at length from multiple models
}
