% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot_compare_recruitment}
\alias{plot_compare_recruitment}
\title{Compare recruitment from multiple models}
\usage{
plot_compare_recruitment(
  object_list,
  object_names,
  figure_dir = "compare_figure/",
  save_plot = TRUE,
  region = NA
)
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{figure_dir}{the directory to save to}

\item{save_plot}{to save the plot to file or not}
}
\description{
Compare recruitment from multiple models
}
