% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{plot_cpue}
\alias{plot_cpue}
\title{Plot CPUE}
\usage{
plot_cpue(
  object,
  scales = "fixed",
  xlab = "Fishing year",
  ylab = "CPUE",
  figure_dir = "figure/"
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot the CPUE data and fit to the data.
}
