% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data-extent.R
\name{plot_data_extent}
\alias{plot_data_extent}
\title{Plot data extent}
\usage{
plot_data_extent(object, xlab = "Fishing year (1 April - 31 March)",
  figure_dir = "figure/", save_plot = TRUE)
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{figure_dir}{the directory to save to}

\item{save_plot}{to save the plot to file or not}
}
\description{
Plot the data extent for tag, LF, CPUE and catch data. Plots data
by year as points where the point size represents the relative
amount of data.
}
