% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pars.R
\name{plot_dens}
\alias{plot_dens}
\title{Plot dens}
\usage{
plot_dens(object, xlab = "Size (mm)", ylab = "Maturation",
  figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot dens
}
