% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth.R
\name{plot_growth_increment}
\alias{plot_growth_increment}
\title{Plot growth increment}
\usage{
plot_growth_increment(
  object,
  xlab = "Size (mm)",
  ylab = "Increment (mm)",
  figure_dir = "figure/",
  empirical = FALSE
)
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}

\item{empirical}{do this or not}
}
\description{
Plot growth increment
}
