% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth.R
\name{plot_growth_matrix}
\alias{plot_growth_matrix}
\title{Plot growth matrix}
\usage{
plot_growth_matrix(object, xlab = "Size after growth (mm)",
  ylab = "Probability", scales = "fixed", figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{scales}{should facet_wrap scales be fixed (the default), free, or free in one dimension (free_x, free_y).}

\item{figure_dir}{the directory to save to}
}
\description{
This function produces two different plots of the growth matrix. The first version is on a single panel. The second version plots a panel for each size category.
}
