% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-handling-mortality.R
\name{plot_handling_mortality}
\alias{plot_handling_mortality}
\title{Plot handling mortality by year}
\usage{
plot_handling_mortality(object, xlab = "Year",
  ylab = "Handling mortality", figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot the fixed handling mortality rates each year.
}
