% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-HR-rapid-updates.R
\name{plot_HR_rapid_updates}
\alias{plot_HR_rapid_updates}
\title{Plot harvest rate for rapid updates 2023}
\usage{
plot_HR_rapid_updates(
  object,
  scales = "free",
  show_mcmc = TRUE,
  xlab = "Fishing year (1 April - 31 March)",
  show_quants = c(0.05, 0.25)
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_mcmc}{show MCMC or not}

\item{xlab}{the x axis label}

\item{show_quants}{the quantiles to plot}

\item{ref}{which reference biomass to plot}
}
\description{
Plot harvest rate for rapid updates 2023
}
