% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-numbers.R
\name{plot_initial_numbers}
\alias{plot_initial_numbers}
\title{Plot initial nubers}
\usage{
plot_initial_numbers(object, xlab = "Size (mm)",
  ylab = "Number of individuals (thousands)", figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot initial nubers
}
