% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-length-frequency.R
\name{plot_lfs}
\alias{plot_lfs}
\title{Length-frequency plots}
\usage{
plot_lfs(
  object,
  n_panel = 10,
  xlab = "Midpoint of size-class (mm)",
  ylab = "Proportion at size (mm)",
  figure_dir = "figure/",
  show_pp = TRUE
)
}
\arguments{
\item{object}{and LSD object}

\item{n_panel}{The number of rows of panels to include per plot.}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}

\item{show_pp}{show the posterior predictive distribution}
}
\description{
Length-frequency plots
}
