% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-length-frequency.R
\name{plot_lfs_resid}
\alias{plot_lfs_resid}
\title{Plot length frequency residuals}
\usage{
plot_lfs_resid(object, figure_dir = "figure/", ylim)
}
\arguments{
\item{object}{and LSD object}

\item{figure_dir}{the directory to save to}

\item{ylim}{the y axis limit}
}
\description{
Plot length frequency residuals
}
