% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-length-frequency.R
\name{plot_lfs_resid2}
\alias{plot_lfs_resid2}
\title{Plot length frequency residuals2}
\usage{
plot_lfs_resid2(object, n_panel = 10, figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object.}

\item{n_panel}{The number of rows of panels to include per plot.}

\item{figure_dir}{the directory to save to.}
}
\description{
Plot length frequency residuals2
}
