% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-osa-resids.R
\name{plot_lfs_resid_OSA}
\alias{plot_lfs_resid_OSA}
\title{Length-frequency OSA residual plots}
\usage{
plot_lfs_resid_OSA(object, figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{figure_dir}{the directory to save to}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}
}
\description{
Length-frequency OSA residual plots
}
