% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-maturation.R
\name{plot_maturation}
\alias{plot_maturation}
\title{Plot maturation}
\usage{
plot_maturation(object, xlab = "Size (mm)", ylab = "Maturation",
  figure_dir = "figure/", empirical = FALSE)
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}

\item{empirical}{plot empirical maturation or not}
}
\description{
Plot maturation
}
