% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mls.R
\name{plot_mls}
\alias{plot_mls}
\title{Plot miminum legal size (mm) by sex and season}
\usage{
plot_mls(object, xlab = "Year", ylab = "Minimum legal size (mm)",
  figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot miminum legal size (mm) by sex and season
}
