% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass.R
\name{plot_money_vuln_biomass}
\alias{plot_money_vuln_biomass}
\title{Plot Money biomass and vulnerable biomass together}
\usage{
plot_money_vuln_biomass(
  object,
  scales = "free",
  show_map = TRUE,
  show_mcmc = TRUE,
  show_proj = FALSE,
  xlab = "Fishing year (1 April - 31 March)",
  show_quants = c(0.05, 0.25)
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_map}{show MAP or not}

\item{show_mcmc}{show MCMC or not}

\item{show_proj}{show projection or not}

\item{xlab}{the x axis label}

\item{show_quants}{the quantiles to plot}

\item{ref}{which reference biomass to plot}
}
\description{
Plot Money biomass and vulnerable biomass together
}
