% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-movement.R
\name{plot_movement}
\alias{plot_movement}
\title{Plot movement}
\usage{
plot_movement(
  object,
  scales = "free",
  show_map = TRUE,
  xlab = "Year",
  ylab = "Movement proportion",
  figure_dir = "figure/"
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_map}{show MAP or not}

\item{xlab}{the x axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot movement
}
