% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mp-rules.R
\name{plot_mp_rules}
\alias{plot_mp_rules}
\title{Plot control rules}
\usage{
plot_mp_rules(mp.pars, cpue.values = numeric(),
  tacc.values = numeric(), cpue = seq(0, 2, 0.01))
}
\arguments{
\item{mp.pars}{a matrix of control parameters with columns as parameters and rows as different rules}

\item{cpue.values}{the current CPUE}

\item{tacc.values}{the current TACC}

\item{cpue}{a vector of the catch per unit effort (CPUE) to be plotted as the x axis}
}
\description{
Plot control rules
}
