% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{plot_offset_cpue}
\alias{plot_offset_cpue}
\title{Plot Offset CPUE}
\usage{
plot_offset_cpue(
  object,
  scales = "fixed",
  xlab = "Offset year (October-September)",
  ylab = "Offset year CPUE (kg/potlift)",
  figure_dir = "figure/"
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot the offset CPUE data and fit to the data.
}
