% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cpue.R
\name{plot_offset_cpue_lm}
\alias{plot_offset_cpue_lm}
\title{Plot offset year CPUE linear model}
\usage{
plot_offset_cpue_lm(object, figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{figure_dir}{the directory to save to}
}
\description{
Plot the offset year CPUE vs the mean of the AW CPUE during year y and SS CPUE during year y-1.
}
