% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-puerulus.R
\name{plot_puerulus}
\alias{plot_puerulus}
\title{Plot puerulus}
\usage{
plot_puerulus(d, scales = "free_y", xlab = "Fishing year",
  ylab = "Puerulus index", figure_dir = "figure/")
}
\arguments{
\item{d}{and LSD object}

\item{scales}{free or fixed}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot puerulus
}
