% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-q.R
\name{plot_q}
\alias{plot_q}
\title{Compare catchability coefficient q from multiple models}
\usage{
plot_q(object, figure_dir = "figure/")
}
\arguments{
\item{object}{as LSD output object}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Compare catchability coefficient q from multiple models
}
