% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot recruitment}
\usage{
plot_recruitment(
  object,
  xlab = "Year",
  ylab = "Recruitment (millions of individuals)",
  figure_dir = "figure/"
)
}
\arguments{
\item{object}{an lsd object}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{figure_dir}{the directory to save the figure to}
}
\description{
This plot shows MAP and posterior of recruitment, MAP and posterior
of R0, and a vertical dashed line indicates the final model year
(after the line is projected recruitment).
}
