% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-recruitment.R
\name{plot_recruitment_size}
\alias{plot_recruitment_size}
\title{Plot recruitment size}
\usage{
plot_recruitment_size(object, xlab = "Size at recruitment (mm)",
  ylab = "Proportion", figure_dir = "figure/")
}
\arguments{
\item{object}{an lsd object}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Plot recruitment size
}
