% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-retro.R
\name{plot_retro}
\alias{plot_retro}
\title{Plot retrospective analysis}
\usage{
plot_retro(object, scales = "free_x", show_map = TRUE,
  figure_dir = "figure/")
}
\arguments{
\item{object}{an LSD object}

\item{scales}{the scales}

\item{show_map}{show the MAP or not}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Plots retrospective analysis for vulnerable biomass and recruitment
}
