% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare-refpoints.R
\name{plot_rules}
\alias{plot_rules}
\title{Plot rules}
\usage{
plot_rules(
  rules,
  rule_labels = NULL,
  fig_name = NULL,
  figure_dir = "compare_figure/",
  width = 8,
  height = 6
)
}
\arguments{
\item{rules}{data frame with rule numbers and parameter values}

\item{rule_labels}{data frame with labels for rules - must be the same length as number of rows in rules}

\item{fig_name}{name for figure if rule labels are included}

\item{figure_dir}{the directory to save the figure to}

\item{width}{option to add width of figure}

\item{height}{option to add height of figure}
}
\description{
Plot rules
}
