% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-selectivity.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity}
\usage{
plot_selectivity(object, xlab = "Size (mm)", figure_dir = "figure/")
}
\arguments{
\item{object}{an S4 object}

\item{xlab}{the x axis label}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Plot selectivity
}
