% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sex-ratio.R
\name{plot_sex_ratio}
\alias{plot_sex_ratio}
\title{Plot sex-ratio}
\usage{
plot_sex_ratio(
  object,
  scales = "free",
  xlab = "Fishing year",
  ylab = "Proportion",
  figure_dir = "figure/"
)
}
\arguments{
\item{object}{an LSD object}

\item{scales}{the scales}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Plot the sex-ratio data and fit to the data.
}
