% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-projections.R
\name{plot_ssb_AW_proj}
\alias{plot_ssb_AW_proj}
\title{Plot AW SSB with projections}
\usage{
plot_ssb_AW_proj(
  object,
  scales = "free",
  show_map = FALSE,
  show_mcmc = TRUE,
  show_proj = TRUE,
  show_ref = TRUE,
  xlab = "Fishing year (1 April - 31 March)"
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_map}{show MAP or not}

\item{show_mcmc}{show MCMC or not}

\item{show_proj}{show projection or not}

\item{show_ref}{show reference level or not}

\item{xlab}{the x axis label}
}
\description{
Plot AW SSB with projections
}
