% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass.R
\name{plot_ssb_recruitment}
\alias{plot_ssb_recruitment}
\title{Plot stock recruit relationship}
\usage{
plot_ssb_recruitment(object, scales = "free_x", show_map = TRUE,
  xlab = "Spawning stock biomass (tonnes)", figure_dir = "figure/")
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_map}{show MAP or not}

\item{xlab}{the x axis label}

\item{figure_dir}{the directory to save to}
}
\description{
Plot stock recruit relationship
}
