% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-surplus-production.R
\name{plot_surplus_production}
\alias{plot_surplus_production}
\title{Plot surplus production}
\usage{
plot_surplus_production(object, xlab = "Fishing year", figure_dir = "figure/")
}
\arguments{
\item{object}{an LSD object}

\item{xlab}{the a axis label}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Plot surplus production
}
