% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth.R
\name{plot_tag_residuals}
\alias{plot_tag_residuals}
\title{Plot tag residuals}
\usage{
plot_tag_residuals(
  object,
  ylab = "Standardised residual",
  figure_dir = "figure/",
  ylim = c(-5, 5)
)
}
\arguments{
\item{object}{and LSD object}

\item{ylab}{the y axis label}

\item{figure_dir}{the directory to save to}

\item{ylim}{the y axis limits}
}
\description{
Plot tag residuals
}
