% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass.R
\name{plot_vulnerable_biomass}
\alias{plot_vulnerable_biomass}
\title{Plot vulnerable reference biomass}
\usage{
plot_vulnerable_biomass(
  object,
  scales = "free_x",
  show_map = TRUE,
  show_mcmc = TRUE,
  show_proj = FALSE,
  xlab = "Fishing year (1 April - 31 March)",
  show_quants = c(0.05, 0.25),
  ref = c("Bmsy", "Bref")
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_map}{show MAP or not}

\item{show_mcmc}{show MCMC or not}

\item{show_proj}{show projection or not}

\item{xlab}{the x axis label}

\item{show_quants}{the quantiles to plot}

\item{ref}{which reference biomass to plot}
}
\description{
Plot vulnerable reference biomass
}
