% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass.R
\name{plot_vulnref_AW}
\alias{plot_vulnref_AW}
\title{Plot AW adjusted vulnerable biomass}
\usage{
plot_vulnref_AW(
  object,
  scales = "free_x",
  show_map = TRUE,
  show_mcmc = TRUE,
  show_proj = FALSE,
  show_ref = FALSE,
  xlab = "Fishing year (1 April - 31 March)"
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_map}{show MAP or not}

\item{show_mcmc}{show MCMC or not}

\item{show_proj}{show projection or not}

\item{show_ref}{show reference level or not}

\item{xlab}{the x axis label}
}
\description{
Plot AW adjusted vulnerable biomass
}
