% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-biomass.R
\name{plot_vulnref_rel}
\alias{plot_vulnref_rel}
\title{Plot relative adjusted vulnerable biomass}
\usage{
plot_vulnref_rel(
  object,
  scales = "free",
  show_proj = TRUE,
  show_map = TRUE,
  show_mcmc = TRUE,
  xlab = "Fishing year",
  figure_dir = "figure/"
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{show_proj}{show projection or not}

\item{show_map}{show MAP or not}

\item{show_mcmc}{show MCMC or not}

\item{xlab}{the x axis label}

\item{show_quants}{the quantiles to plot}

\item{ref}{which reference biomass to plot}
}
\description{
Plot relative adjusted vulnerable biomass
}
