% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare-refpoints.R
\name{read_catch}
\alias{read_catch}
\title{Read Catch and CPUE info}
\usage{
read_catch(object_list, object_names)
}
\arguments{
\item{object_list}{list of mcmc results}

\item{object_names}{list of model names}
}
\description{
Read Catch and CPUE info
}
