% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare-refpoints.R
\name{read_SSB}
\alias{read_SSB}
\title{Read SSB info}
\usage{
read_SSB(object_list, object_names)
}
\arguments{
\item{object_list}{list of mcmc results}

\item{object_names}{list of model names}
}
\description{
Read SSB info
}
