% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-stan-mcmc.R
\name{read_stan_mcmc}
\alias{read_stan_mcmc}
\title{Read Stan MCMC}
\usage{
read_stan_mcmc(mcmcfiles)
}
\arguments{
\item{mcmcfiles}{character vector of the complete path to csv output files produced by \code{cmdStan}}
}
\description{
Read MCMC comma delimited files produced by \code{cmdStan}
}
