% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsdOutput-class.R
\docType{methods}
\name{show,lsdOutput-method}
\alias{show,lsdOutput-method}
\title{Dunno what this does}
\usage{
\S4method{show}{lsdOutput}(object)
}
\arguments{
\item{object}{an LSD object}
}
\description{
Dunno what this does
}
