% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{table_compare_residuals}
\alias{table_compare_residuals}
\title{Table comparing residuals for various data types across models}
\usage{
table_compare_residuals(object_list, object_names,
  figure_dir = "compare_figure/")
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{figure_dir}{the directory to save to}
}
\description{
Table comparing residuals for various data types across models
}
