% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-parameters.R
\name{table_parameters}
\alias{table_parameters}
\title{Create residuals table}
\usage{
table_parameters(object, figure_dir = "figure/", save_table = TRUE)
}
\arguments{
\item{object}{and LSD object}

\item{figure_dir}{the directory to save to}

\item{save_table}{TRUE or FALSE}
}
\description{
Create residuals table
}
