% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-residuals.R
\name{table_residuals}
\alias{table_residuals}
\title{Create residuals table}
\usage{
table_residuals(object, figure_dir = "figure/", save_table)
}
\arguments{
\item{object}{and LSD object}

\item{figure_dir}{the directory to save to}

\item{save_table}{TRUE or FALSE}
}
\description{
Create residuals table
}
