% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TACC.R
\name{TACC}
\alias{TACC}
\title{Calculate the TACC given a vector of CPUE}
\usage{
TACC(par, cpue)
}
\arguments{
\item{par}{the parameters of the rule}

\item{cpue}{the CPUE}
}
\description{
Calculate the TACC given a vector of CPUE
}
