% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-lsd.R
\name{theme_lsd}
\alias{theme_lsd}
\title{LSD theme based on ggplot2 theme}
\usage{
theme_lsd(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
LSD theme based on ggplot2 theme
}
