% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-model-code.R
\name{parse_model_code}
\alias{parse_model_code}
\title{Parse model code}
\usage{
parse_model_code(base = "base_lsd", funcs = c("fun_curves",
  "fun_growth_increment", "fun_growth_matrix", "fun_length_weight",
  "fun_tail_compression", "fun_equilibrium_numbers",
  "fun_recruitment_size", "fun_fishing_mortality", "fun_MAR", "fun_MSY"),
  ctl = "lsd.ctl", save = FALSE)
}
\arguments{
\item{base}{the core model code}

\item{funcs}{the functions that are called by the core model code}

\item{ctl}{the .ctl file that will be input to the model. This is needed for switching parameters on/off.}

\item{save}{should the full model be written to file or not}
}
\description{
This reads the core model code and combines it with the functions used by the model.
}
