% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compare.R
\name{plot_compare_lfs}
\alias{plot_compare_lfs}
\title{Compare LFs from multiple models}
\usage{
plot_compare_lfs(
  object_list,
  object_names,
  yrs = 2018:2019,
  xlab = "Midpoint of size-class (mm)",
  ylab = "Proportion at size (mm)"
)
}
\arguments{
\item{object_list}{list of 'lsd.rds' files from multiple models}

\item{object_names}{vector of model names associated with each of the output files in object_list}

\item{yrs}{the years to compare}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}
}
\description{
Compare LFs from multiple models
}
