% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-F.R
\name{plot_F}
\alias{plot_F}
\title{Plot fishing mortality}
\usage{
plot_F(
  object,
  scales = "free",
  xlab = "Fishing year",
  figure_dir = "figure/",
  show_proj = FALSE
)
}
\arguments{
\item{object}{and LSD object}

\item{scales}{free or fixed}

\item{xlab}{the x axis label}

\item{figure_dir}{the directory to save to}

\item{show_proj}{show projection or not}
}
\description{
Plot fishing mortality
}
