% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-refpoints.R
\name{plot_refpoints}
\alias{plot_refpoints}
\title{Calculate reference points and create relevant figures}
\usage{
plot_refpoints(object, figure_dir)
}
\arguments{
\item{object}{the ref-lsd object}

\item{figure_dir}{the directory to save figures to}
}
\description{
Calculate reference points and create relevant figures
}
