% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-snail.R
\name{plot_snail}
\alias{plot_snail}
\title{Plot snail trial}
\usage{
plot_snail(object, figure_dir = "figure/")
}
\arguments{
\item{object}{an LSD object}

\item{figure_dir}{the directory to save the figure to}
}
\description{
Plot snail trial
}
